backend/requirements.txt
# BrandMeld Backend Dependencies

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6

# Google AI — unified new SDK only
google-genai>=1.0.0

# JWT validation for Supabase auth middleware
PyJWT>=2.8.0

# Database
supabase>=2.0.0

# Web Automation
playwright>=1.40.0

# Utilities
python-dotenv==1.0.1
